requirements.txt
attrs==22.1.0
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
lxml==4.9.1
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
pycodestyle==2.9.1
pyparsing==3.0.9
pytest==7.2.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2022.4
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
XlsxWriter==3.0.3
